requirements.txt
dash
pandas
plotly
pyyaml
redis
websockets
websocket-client
